requirements-gpu.txt
tensorflow-gpu
matplotlib
numpy
shutil
glob
tqdm